Some More Improvements to Speed
Turns out there was some totally useless stuff I was doing upon loading a folder.
Since v1.1 limits the number of accounts you can see in the list to 25, it would be totally useless to loop through the rest of the files when we didn't do anything really useful with them besides incrementing the progress bar's value. Which I did in one simple step.